refactor: shadcn init
