+++ package.json
@@ -43,16 +43,21 @@
     "@types/react-dom": "^19.1.2",
     "@types/semver": "^7.7.0",
     "@vitejs/plugin-react": "^4.4.1",
+    "class-variance-authority": "^0.7.1",
+    "clsx": "^2.1.1",
     "electron": "^35.2.1",
     "electron-builder": "^26.0.12",
     "electron-vite": "^3.1.0",
+    "lucide-react": "^0.555.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-hook-form": "^7.56.1",
     "react-i18next": "^15.5.1",
     "react-router-dom": "^7.5.2",
     "swr": "^2.3.3",
+    "tailwind-merge": "^3.4.0",
     "tailwindcss": "^4.1.17",
+    "tw-animate-css": "^1.4.0",
     "type-fest": "^4.40.1",
     "typescript": "^5.8.3",
     "vite": "^6.3.3"
