+++ src/renderer/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta http-equiv="Content-Security-Policy" content="script-src 'self'; worker-src blob:;" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link href="/src/styles/global.css" rel="stylesheet" />
+    <link href="/src/styles/globals.css" rel="stylesheet" />
     <title>electron-nestjs-react-vite-template</title>
   </head>
   <body>
