--- src/renderer/src/styles/global.css
@@ -1,1 +1,0 @@
-@import 'tailwindcss';
