+++ src/renderer/src/styles/globals.css
@@ -1,0 +1,4 @@
+@import 'tailwindcss';
+@import 'tw-animate-css';
+
+@import './shadcn-theme.css';
